+++ center/router/router_notify_channel.go
@@ -39,7 +39,7 @@ func (rt *Router) notifyChannelsAdd(c *gin.Context) {
 	}
 
 	ginx.Dangerous(models.DB(rt.Ctx).CreateInBatches(lst, 100).Error)
-	ids := make([]uint, 0, len(lst))
+	ids := make([]int64, 0, len(lst))
 	for _, tpl := range lst {
 		ids = append(ids, tpl.ID)
 	}
